+++ unique_toolkit/unique_toolkit/tools/tool_manager.py
@@ -13,7 +13,7 @@ from unique_toolkit.tools.schemas import ToolCallResponse, ToolPrompts
 from unique_toolkit.tools.tool import Tool
 from unique_toolkit.tools.tool_progress_reporter import ToolProgressReporter
 from unique_toolkit.tools.utils.execution.execution import Result, SafeTaskExecutor
-from unique_toolkit.unique_toolkit.evals.schemas import EvaluationMetricName
+from unique_toolkit.evals.schemas import EvaluationMetricName
 
 
 class ForcedToolOption: