+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.12] - 2025-08-18
+- Adding initial versions of
+  - Evaluation Manager
+  - History Manager
+  - Postprocessor Manager
+  - Thinking Manager
+- Updated tool manager
+
 ## [0.8.11] - 2025-08-15
 - Fix no tool call respoonse in ChatMessage -> Open Ai messages translation
 - Add simple append method to OpenAIMessageBuilder